+++ boards/TMC8462_eval.c
@@ -32,8 +32,8 @@ void TMC8462_init_ch2(void);
 static void periodicJob(uint32_t actualSystick);
 static void register_write(uint8_t motor, uint16_t address, int32_t value);
 static void register_read(uint8_t motor, uint16_t address, int32_t *value);
-static void memory_read(uint8_t motor, uint8_t address, int32_t *value);
-static void memory_write(uint8_t motor, uint8_t address, int32_t value);
+static void memory_read(uint8_t motor, uint16_t address, int32_t *value);
+static void memory_write(uint8_t motor, uint16_t address, int32_t value);
 static void pdi_reset(void);
 static uint32_t eep_read(int32_t *value);
 static uint32_t eep_write(int32_t value);
@@ -139,12 +139,12 @@ static void register_read(uint8_t motor, uint16_t address, int32_t *value)
 	}
 }
 
-static void memory_read(uint8_t motor, uint8_t address, int32_t *value)
+static void memory_read(uint8_t motor, uint16_t address, int32_t *value)
 {
 	*value = tmc8462_esc_read_16(&TMC8462, (motor << 8) | address);
 }
 
-static void memory_write(uint8_t motor, uint8_t address, int32_t value)
+static void memory_write(uint8_t motor, uint16_t address, int32_t value)
 {
 	tmc8462_esc_write_8(&TMC8462, (motor << 8) | address, BYTE(value, 0));
 }
